style: rename cms paragraph schema.ts to validation.ts
