+++ src/actions/cms/paragraphs/create.ts
@@ -1,9 +1,9 @@
 'use server'
-import { createCmsParagraphActionValidation } from './schema'
+import { createCmsParagraphActionValidation } from './validation'
 import { createCmsParagraph } from '@/server/cms/paragraphs/create'
 import { safeServerCall } from '@/actions/safeServerCall'
 import { createZodActionError } from '@/actions/error'
-import type { CreateCmsParagraphActionTypes } from './schema'
+import type { CreateCmsParagraphActionTypes } from './validation'
 import type { CmsParagraph } from '@prisma/client'
 import type { ActionReturn } from '@/actions/Types'
